pyproject.toml
[project]
name = "blockchain-voice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "elevenlabs>=1.6.0",
  "python-dotenv>=1.0.1",
  "requests>=2.32.3",
  "sounddevice>=0.4.7",
  "soundfile>=0.12.1",
]
